+++ resources/attributes/default.rb
@@ -80,7 +80,7 @@ default['redborder']['memory_services']['http2k'] = { 'count': 10, 'memory': 0 }
 default['redborder']['memory_services']['kafka'] = { 'count': 120, 'memory': 0, 'max_limit': 524288 }
 default['redborder']['memory_services']['n2klocd'] = { 'count': 10, 'memory': 0 }
 default['redborder']['memory_services']['postgresql'] = { 'count': 25, 'memory': 0 }
-default['redborder']['memory_services']['aerospike'] = { 'count': 10, 'memory': 0 }
+default['redborder']['memory_services']['aerospike'] = { 'count': 50, 'memory': 0 }
 default['redborder']['memory_services']['rb-aioutliers'] = { 'count': 10, 'memory': 0 }
 default['redborder']['memory_services']['redborder-agents'] = { 'count': 5, 'memory': 0 }
 default['redborder']['memory_services']['redborder-cep'] = { 'count': 80, 'memory': 0 }
@@ -93,6 +93,7 @@ default['redborder']['memory_services']['secor-vault'] = { 'count': 30, 'memory'
 default['redborder']['memory_services']['redis'] = { 'count': 10, 'memory': 0 }
 default['redborder']['memory_services']['airflow-scheduler'] = { 'count': 30, 'memory': 0 }
 default['redborder']['memory_services']['airflow-webserver'] = { 'count': 30, 'memory': 0 }
+default['redborder']['memory_services']['rb-reputation'] = { 'count': 30, 'memory': 0 }
 
 # default attributes for managers_info, it would be rewriten with the cluster config
 default['redborder']['cluster_info'] = {}
@@ -176,6 +177,7 @@ default['redborder']['services']['secor-vault']               = false
 default['redborder']['services']['redis']                     = false
 default['redborder']['services']['airflow-scheduler']         = false
 default['redborder']['services']['airflow-webserver']         = false
+default['redborder']['services']['rb-reputation']             = false
 
 default['redborder']['systemdservices']['aerospike']                = ['aerospike']
 default['redborder']['systemdservices']['chef-client']              = ['chef-client']
@@ -226,15 +228,30 @@ default['redborder']['systemdservices']['secor-vault']              = ['rb-secor
 default['redborder']['systemdservices']['redis']                    = ['redis']
 default['redborder']['systemdservices']['airflow-scheduler']        = ['airflow-scheduler']
 default['redborder']['systemdservices']['airflow-webserver']        = ['airflow-webserver']
+default['redborder']['systemdservices']['rb-reputation']            = ['rb-reputation']
 
-default['redborder']['manager']['balanced'] = [ { port: 443, protocol: 'tcp', name: 'redborder webui', service: 'webui', redirected_service: 'nginx', persistence_timeout: 9600 }, { port: 2055, protocol: 'udp', name: 'netflow,ipfix/sflow daemon', service: 'f2k', redirected_service: 'f2k', persistence_timeout: 30 }, { port: 6343, protocol: 'udp', name: 'sflow daemon', service: 'sfacctd', redirected_service: 'sfacctd', persistence_timeout: 30 }, { port: 9092, protocol: 'tcp', name: 'kafka', service: 'kafka', redirected_service: 'kafka', persistence_timeout: 30 } ]
+# Balanced services
+default['redborder']['manager']['balanced'] = [
+  { port: 443, protocol: 'tcp', name: 'redborder nginx', service: 'nginx', redirected_service: 'nginx', persistence_timeout: 9600 },
+  { port: 2055, protocol: 'udp', name: 'netflow,ipfix/sflow daemon', service: 'f2k', redirected_service: 'f2k', persistence_timeout: 30 },
+  { port: 6343, protocol: 'udp', name: 'sflow daemon', service: 'sfacctd', redirected_service: 'sfacctd', persistence_timeout: 30 },
+  { port: 9092, protocol: 'tcp', name: 'kafka', service: 'kafka', redirected_service: 'kafka', persistence_timeout: 30 },
+]
 
 # Tier
 default['redborder']['druid']['historical']['tier'] = 'default'
 default['redborder']['druid']['historical']['maxsize'] = -1
 
 # Virtual Ips
-default['redborder']['manager']['virtual_ips'] = { internal: [{ service: 'postgresql' }], external: [{ service: 'webui', deps: ['nginx'] }, { service: 'f2k' }, { service: 'sfacctd' }, { service: 'kafka' }] }
+default['redborder']['manager']['virtual_ips'] = {
+  internal: [{ service: 'postgresql' }],
+  external: [
+    { service: 'nginx' },
+    { service: 'f2k' },
+    { service: 'sfacctd' },
+    { service: 'kafka' },
+  ],
+}
 
 default['redborder']['pending_changes'] = 0
 
@@ -244,3 +261,79 @@ default['redborder']['vault_incidents_priority_filter'] = 'error'
 
 # Save Secor S3 raw path
 default['redborder']['manager']['s3rawpath'] = 'rbraw'
+
+# rb-reputation settings
+default['redborder']['manager']['rb-reputation'] = {}
+default['redborder']['manager']['rb-reputation']['workers'] = 300
+
+# Policy Enforced
+default['redborder']['manager']['policy_enforced'] = {}
+default['redborder']['manager']['policy_enforced']['production'] = {}
+default['redborder']['manager']['policy_enforced']['production']['host'] = ''
+default['redborder']['manager']['policy_enforced']['production']['port'] = '8080'
+default['redborder']['manager']['policy_enforced']['production']['actions'] = {}
+default['redborder']['manager']['policy_enforced']['production']['actions']['delete_computer'] = {}
+default['redborder']['manager']['policy_enforced']['production']['actions']['delete_computer']['url'] = '/wm/acl/rules/json'
+default['redborder']['manager']['policy_enforced']['production']['actions']['delete_computer']['verb'] = 'Delete'
+default['redborder']['manager']['policy_enforced']['production']['actions']['delete_all'] = {}
+default['redborder']['manager']['policy_enforced']['production']['actions']['delete_all']['url'] = '/wm/acl/clear/json'
+default['redborder']['manager']['policy_enforced']['production']['actions']['delete_all']['verb'] = 'Get'
+default['redborder']['manager']['policy_enforced']['production']['actions']['list_computers'] = {}
+default['redborder']['manager']['policy_enforced']['production']['actions']['list_computers']['url'] = '/wm/acl/rules/json'
+default['redborder']['manager']['policy_enforced']['production']['actions']['list_computers']['verb'] = 'Get'
+
+default['redborder']['manager']['policy_enforced']['development'] = {}
+default['redborder']['manager']['policy_enforced']['development']['host'] = ''
+default['redborder']['manager']['policy_enforced']['development']['port'] = '8080'
+default['redborder']['manager']['policy_enforced']['development']['actions'] = {}
+default['redborder']['manager']['policy_enforced']['development']['actions']['delete_computer'] = {}
+default['redborder']['manager']['policy_enforced']['development']['actions']['delete_computer']['url'] = '/wm/acl/rules/json'
+default['redborder']['manager']['policy_enforced']['development']['actions']['delete_computer']['verb'] = 'Delete'
+default['redborder']['manager']['policy_enforced']['development']['actions']['delete_all'] = {}
+default['redborder']['manager']['policy_enforced']['development']['actions']['delete_all']['url'] = '/wm/acl/clear/json'
+default['redborder']['manager']['policy_enforced']['development']['actions']['delete_all']['verb'] = 'Get'
+default['redborder']['manager']['policy_enforced']['development']['actions']['list_computers'] = {}
+default['redborder']['manager']['policy_enforced']['development']['actions']['list_computers']['url'] = '/wm/acl/rules/json'
+default['redborder']['manager']['policy_enforced']['development']['actions']['list_computers']['verb'] = 'Get'
+
+# Default malware loaders and scores
+default['redborder']['manager']['loaders']                                    = {}
+default['redborder']['manager']['loaders']['fuzzy']                           = {}
+default['redborder']['manager']['loaders']['fuzzy']['ssdeep_oldhashes_limit']     = 1000
+default['redborder']['manager']['loaders']['fuzzy']['ssdeep_malwarehashes_limit'] = 1000
+
+default['redborder']['manager']['malware'] = {}
+default['redborder']['manager']['malware']['threshold'] = 50
+
+default['redborder']['manager']['loaders']['scores']                          = {}
+default['redborder']['manager']['loaders']['scores']['hash']                  = {}
+default['redborder']['manager']['loaders']['scores']['hash']['g_rt_score']    = 1
+default['redborder']['manager']['loaders']['scores']['hash']['g_fb_score']    = 0
+default['redborder']['manager']['loaders']['scores']['hash']['g_sb_score']    = 0
+default['redborder']['manager']['loaders']['scores']['hash']['rt_virustotal'] = 0.5
+default['redborder']['manager']['loaders']['scores']['hash']['rt_metascan']   = 0.5
+# default['redborder']['manager']['loaders']['scores']['hash']['rt_brightcloud'] = 0
+default['redborder']['manager']['loaders']['scores']['hash']['fb_virustotal'] = 0.4
+default['redborder']['manager']['loaders']['scores']['hash']['fb_metascan']   = 0.25
+default['redborder']['manager']['loaders']['scores']['hash']['fb_clamscan']   = 0.25
+# default['redborder']['manager']['loaders']['scores']['hash']['fb_decompress'] = 0
+default['redborder']['manager']['loaders']['scores']['hash']['fb_yara']       = 0.1
+# default['redborder']['manager']['loaders']['scores']['hash']['fb_kaspersky']  = 0
+default['redborder']['manager']['loaders']['scores']['hash']['sb_fuzzy']      = 0.5
+default['redborder']['manager']['loaders']['scores']['hash']['sb_cuckoo']     = 0.5
+
+default['redborder']['manager']['loaders']['scores']['ip'] = {}
+default['redborder']['manager']['loaders']['scores']['ip']['g_rt_score']    = 1
+default['redborder']['manager']['loaders']['scores']['ip']['g_fb_score']    = 0
+default['redborder']['manager']['loaders']['scores']['ip']['g_sb_score']    = 0
+default['redborder']['manager']['loaders']['scores']['ip']['rt_otx']        = 0.5
+default['redborder']['manager']['loaders']['scores']['ip']['rt_metascan']   = 0.5
+# default['redborder']['manager']['loaders']['scores']['ip']['rt_brightcloud'] = 0
+
+default['redborder']['manager']['loaders']['scores']['url'] = {}
+default['redborder']['manager']['loaders']['scores']['url']['g_rt_score']      = 1
+default['redborder']['manager']['loaders']['scores']['url']['g_fb_score']      = 0
+default['redborder']['manager']['loaders']['scores']['url']['g_sb_score']      = 0
+# default['redborder']['manager']['loaders']['scores']['url']['rt_brightcloud']  = 0
+default['redborder']['manager']['loaders']['scores']['url']['rt_metascan']     = 0.5
+default['redborder']['manager']['loaders']['scores']['url']['rt_virustotal']   = 0.5